requirements.txt
# FileCherry Python Dependencies
#
# Core framework
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.5.0

# HTTP clients
httpx>=0.25.0

# Utilities
rich>=13.7.0
pyyaml>=6.0.0

# Document processing
pypdf>=3.17.0
python-docx>=1.1.0
beautifulsoup4>=4.12.0
lxml>=4.9.0

# Embeddings and vector search
sentence-transformers>=2.2.0
numpy>=1.24.0

# Testing
pytest>=7.4.0
pytest-mock>=3.12.0
pytest-cov>=4.1.0

# Development tools
black>=23.11.0
ruff>=0.1.6
mypy>=1.7.0

